# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.AcceptFriendRequest:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
...
  requests.DeclineFriendRequest:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive social media backend with posts, video calling, and
    file uploads
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Social Media Backend API
  version: "2.0"
paths:
...
  /friends:
    get:
      description: Get list of user's friends with pagination
      parameters:
      - default: 20
        description: Number of friends to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: before
        type: string
      - description: Cursor for pagination
        in: query
        name: after
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Friends list
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get friends list
      tags:
      - Friends
  /friends/{id}:
    delete:
      description: Remove a user from friends list
      parameters:
      - description: Friend user ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Friend removed
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid friend ID or not friends
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove friend
      tags:
      - Friends
  /friends/{id}/block:
    post:
      description: Block a user and remove from friends if applicable
      parameters:
      - description: User ID to block
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User blocked
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID or already blocked
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Block user
      tags:
      - Friends
  /friends/{id}/status:
    get:
      description: Check friendship status between current user and another user
      parameters:
      - description: User ID to check
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Friendship status
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Check friendship status
      tags:
      - Friends
  /friends/{id}/unblock:
    post:
      description: Unblock a previously blocked user
      parameters:
      - description: User ID to unblock
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User unblocked
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID or not blocked
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Unblock user
      tags:
      - Friends
  /friends/accept-request:
    post:
      description: Accept a friend request from another user
      parameters:
      - description: Sender user ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.AcceptFriendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Friend request accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID or request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Request not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Accept friend request
      tags:
      - Friends
  /friends/blocked:
    get:
      description: Get list of users blocked by current user
      parameters:
      - default: 20
        description: Number of users to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Blocked users list
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get blocked users
      tags:
      - Friends
  /friends/decline-request:
    post:
      description: Decline a friend request from another user
      parameters:
      - description: Sender user ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.DeclineFriendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Friend request declined
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID or request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Request not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Decline friend request
      tags:
      - Friends
  /friends/requests:
    get:
      description: Get list of received or sent friend requests
      parameters:
      - default: '"all"'
        description: 'Request type: received, sent, or all'
        enum:
        - all
        - sent
        - received
        in: query
        name: type
        required: true
        type: string
      - default: 5
        description: Number of requests to return
        in: query
        name: limit
        required: true
        type: integer
      - description: Cursor for pagination
        in: query
        name: before
        type: string
      - description: Cursor for pagination
        in: query
        name: after
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Friend requests list
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get friend requests
      tags:
      - Friends
  /friends/requests-stats:
    get:
      description: Get stats of received or sent friend requests
      produces:
      - application/json
      responses:
        "200":
          description: Friend requests stats
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get friend requests stats
      tags:
      - Friends
  /friends/send-request:
    post:
      consumes:
      - application/json
      description: Send a friend request to another user
      parameters:
      - description: Target user ID
        in: path
        name: id
        required: true
        type: integer
      - description: Optional message
        in: body
        name: request
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Friend request sent
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or user ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Send friend request
      tags:
      - Friends
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...